# backend/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Trello Clone API
  description: |
    REST API for the Trello Clone app.

    **Authentication:** Most endpoints require `Authorization: Bearer <access_token>`.
    Workspace/board endpoints also require a verified email.

    **WebSockets** (not covered by this spec):
    - `GET /ws?board_id={uuid}&token={access_token}` — board updates
    - `GET /ws/notifications?token={access_token}` — user notifications
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://trello-api-8oul.onrender.com/api/v1
    description: Production (Render)
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Email not verified
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Generic success (may include verification_url when email is off)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationEmailResponse"
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Generic success (may include reset_url when email is off)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetEmailResponse"
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    RegisterResponse:
      allOf:
        - $ref: "#/components/schemas/AuthResponse"
        - type: object
          properties:
            verification_sent:
              type: boolean
            email_verified:
              type: boolean
            verification_url:
              type: string
              description: Present when email was not sent (dev / no mailer)
    VerificationEmailResponse:
      type: object
      properties:
        message:
          type: string
        verification_url:
          type: string
    PasswordResetEmailResponse:
      type: object
      properties:
        message:
          type: string
        reset_url:
          type: string
...